% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whereIsThisPackage.R
\name{whereIsThisPackage}
\alias{whereIsThisPackage}
\title{Which repository do my R packages belong to?}
\usage{
whereIsThisPackage(list.of.pkgs)
}
\arguments{
\item{list.of.pkgs}{list of R package names}
}
\value{
A list containing two sub lists corresponding to CRAN or Bioconductor
}
\description{
This function accepts a list of R package names and groups them in two separate
lists corresponding to CRAN or Bioconductor. It also list the package names
that do not match with the package names in either repositories.
}
